move to snapping/session.tsx
