+++ src/app/views/canvas/snapping/session.tsx
@@ -1,26 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
-import { Specification, Prototypes } from "../../../../core";
+import { Prototypes } from "../../../../core";
 import { HandlesDragEvent } from "../handles/common";
-
-export interface SnappableGuide<ElementType> {
-  element: ElementType;
-  guide: Prototypes.SnappingGuides.Description;
-}
-
-export interface SnappingAction<ElementType> {
-  type: "snap" | "move" | "property" | "value-mapping";
-  attribute?: string;
-  property?: string;
-  field?: string | string[];
-
-  // move
-  value?: Specification.AttributeValue;
-
-  // snap
-  snapElement?: ElementType;
-  snapAttribute?: string;
-}
+import { SnappableGuide, SnappingAction } from "./common";
 
 export class SnappingSession<ElementType> {
   public candidates: Array<SnappableGuide<ElementType>>;
